requirements.txt
django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
factory-boy==3.3.3
faker==37.1.0
flake8==7.2.0
isort==6.0.1
pydantic==2.11.1
pytest==8.3.5
pytest-django==4.11.1
pytest-mock==3.14.0
requests==2.32.3
